MB-55978 Use builder flag to indicate join planning
Currently for join planning we rely on a couple of flags on
the KeyspaceTerm to indicate whether a nested-loop join or
a hash join is being planned. While this works it is a bit
cumbersome, for example during join enumeration we need to
remember not to permenantly mark the KeyspaceTerms since
we'll be considering multiple plans, instead once the best
plan is chosen we need to walk the plan and mark the
KeyspaceTerms with the appropriate flag.

The previous fix for this bug deals with ExpressionTerm,
and it uses a flag on the builder itself to determine
whether nested-loop is being considered, and it put the
marker on the plan operator itself rather than the
KeyspaceTerm. Since the plan operator is "final" once it
is generated, we don't need to keep unsetting/resetting
of the flag. (it is done this way since we don't already
have similar flags on ExpressionTerm, only on KeyspaceTerm
so far).

This scheme works better than what we currently have, and
thus we change to use this scheme for KeyspaceTerm as well.
Instead of setting/unsetting the flags on the KeyspaceTerm,
we use the flag on the builder itself and set the
appropriate flag on the plan operator. Now both nested-loop
join and hash join are handled this way.

Change-Id: I5cd82bbf791d646a5903eb757f39f78a168d039f
Reviewed-on: https://review.couchbase.org/c/query/+/188400
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>