MB-56040 Change FFDC retention strategy
Changes the retention strategy used for FFDC to maintain a history by
"event" and with each event's history, retain the oldest and newest
5 occurrences.

Also detects external file removal (when it needs to prune the history)
allowing external management too.

Each event can be triggered twice (must be at least 10 seconds apart)
before a reset is necessary. A reset enables gathering to resume for an
event and is done when the trigger conditions are not met. (e.g. once
the queue is no longer full)

Change-Id: If442640e4bcca78e9c8c49e27a82565d5c4893d1
Reviewed-on: https://review.couchbase.org/c/query/+/188541
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>