MB-56077 Better generation of index filter in case of OR predicates
After the fix for MB-56039, there are more situations where we do not
use individual filters for sarging indexes when OR predicates is present,
we use the complete DNF-transformed OR clause instead.

For generation of index filters, when using individual filters, we keep
track of the filters that generates exact index spans, and do not
include these filters as index filters, since exact spans implies that
the filter is already applied from the index scan. However this
mechanism does not work if we do not use individual filters. This
results in filters appear as index filters unnecessarily, and can
mess up selectivity calculation when CBO is in use.

The fix includes the following:
 - add a flag in indexEntry such that when we use individual filters
   for OR predicates we remember that
 - for OR predicates if we did not use individual filters, add a new
   function to try to derive index filter from each individual filter,
   we do that by trying each individual index key, and "remove" the
   part of the original filter that can be used to generate an exact
   index span. This way we don't generate unnecessary index filters.
 - no need to generate index filters if _PUSHDOWN_EXACTSPANS flag is
   present (all filters pushed to index scan)
 - modify the fix for MB-42474, such that instead of setting exact
   to false, we use another new flag on the indexEntry. This preserves
   the meaning of "exact" while still detect and prevent the complete
   index pushdown from that defect (also OR-prediate related).

Change-Id: I6e5ce67c67e550eda6539b824b83d356c819fa25
Reviewed-on: https://review.couchbase.org/c/query/+/188575
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>