MB-56088 Improve ADVISE for queries with multiple levels of UNNEST
The problematic query is of the form:

SELECCT ... FROM default d UNNEST d.arr1 AS a UNNEST a.arr2 AS b WHERE d.c1 = 1

There are two levels of unnests, with unnest aliases 'a' and 'b',
however there are no predicate in the WHERE clause on either of the
unnest aliases.

The current index advisor code assumes that in this case there is
some predicates on the unnest aliases, and when there are none, it
end up generating a string with three consecutive connectors (_)
which is unexpected and leads to issues later on.

When a query with unnests is planned, we internally derive a
IS NOT MISSING predicate on the (inner) unnest alias, to facilitate
index selection. However this derivation only occurs for what's
considered "primary unnest", which are the level-1 unnests on the
primary term. In case of multiple levels of unnest like above, we
currently do not derive the predicate on level-2 unnest alias ('b'
above), and if an index is defined on two levels of ALL ARRAY
(ALL ARRAY (ALL ARRAY b FOR b IN a.arr2 END) FOR a IN arr1 END),
this index cannot be used currently since we are missing a predicate
on the innermost array key ('b'). We do derive a predicate on unnest
alias 'a', but this is not useful for index selection for an index
key with nested array keys.

The index advisor code disregard the derived IS NOT MISSING
predicate (since it's not user-specified).

The fix for this defect includes the following aspects:

 - when deriving IS NOT MISSING predicates on unnest aliases,
   include all (inner) unnest aliases, since we don't know
   what kind of indexes are present (could be on level-1 or
   level-2 unnest aliases).
 - in the index advisor code, if we end up with no predicates
   on any unnest aliases, use the unnest alias on the highest
   level unnest, assuming there will be a IS NOT MISSING
   predicate on this unnest alias
 - with multiple levels of unnest, if predicate exists on a
   subset of the unnest aliases, generate an index key that
   corresponds to that unnest alias, instead of always using
   the highest level of unnest. E.g.,
      UNNEST d.arr1 AS a UNNEST a.arr2 AS b WHERE a.x = 10
   in this case since there is a predicate on 'a', but no
   predicate on 'b', generate the array index key:
      ALL ARRAY a.x FOR a IN arr1 END
   i.e., on level-1 unnest only, not on nested unnest
 - set the unnest flag in queryInfo only with inner unnests,
   since array index cannot be used for outer unnests

Change-Id: Id78789cfd7c40f7d23eaf6301bbc425b124972a6
Reviewed-on: https://review.couchbase.org/c/query/+/188635
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>