MB-56104 Add checking for index hint errors when comparing single index scan vs union scan
If query has OR-clause as predicate, planner could build both a
single index plan and an UnionScan plan. If both plans are
available, the planner picks one with a variety of criteria,
however currently this does not include checking for any
optimizer hint error.

Add checking for optimizer hint errors and pick the one without
hint error if available.

Change-Id: I1f4b3bdd92cb3ebb30ec9c0160b3e810a5cbd8a5
Reviewed-on: https://review.couchbase.org/c/query/+/188673
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>