MB-56112 Pick covering index with more consecutive sargable keys
When two covering indexes are being compared, we currently
check for sumKeys and total index length, for two similar
indexes that have the same set of index keys in different
ordering, the current check does not differentiate the
two indexes, and pick randomly between the two.

Add a check for minKeys as well, similar to the existing
check for sumKeys. In case the two indexes are similar,
pick the one with more consecutive sargable keys at the
beginning of the index keys (more effective index search).

Change-Id: I43073fb587c6c43842dbec9ea096a669a9ceaa59
Reviewed-on: https://review.couchbase.org/c/query/+/188694
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>