MB-56117 Use inner unnest to perform ANSI outer join to inner join transformation
For a query with left-outer join, we perform outer join to
inner join transformation based on available predicates.

If there is inner unnest on an array from the subservient
side of the keyspace, then we should be able to convert
the subservient keyspace to level 0, since any
missing-extended result should be thrown out at the
(inner) unnest step.

E.g.:

SELECT ... FROM t1 LEFT JOIN t2 ON t1.a = t2.a UNNEST t2.arr WHERE ...

Current logic adds the unnest alias to a list of aliases
to be checked via other predicates. This is not necessary.
The presence of the inner unnest itself is enough.

Change-Id: Ia24c9513877cc762ed3f8199f25ac6d21f91c043
Reviewed-on: https://review.couchbase.org/c/query/+/188701
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>