MB-56068 Add support for GOQUERY_GOMAXPROCS
This adds support for the GOQUERY_GOMAXPROCS environment variable to
allow a Query-specific GOMAXPROCS setting.

The max-procs settings is always the lower of the number of logical CPUs
(CE limited) and evironment variable setting.  If GOQUERY_GOMAXPROCS is
set then it is the only variable consulted; if not then GOMAXPROCS is
used.  If an environment variable is set to an invalid value it is
considered set but the value is ignored.

Change-Id: I61a603e3a7adc36bc9e2338802cb5665eca1f4e5
Reviewed-on: https://review.couchbase.org/c/query/+/188802
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>