MB-56372 : Add profile timings for embedded N1QL statements in UDFs
Profile information for these embedded statements will be in an array "udfStatements" in the profile section of the response

"profile": {
    "executionTimings": {
        "udfStatements": [
            ...
        ]
    }
}

Store all execution trees and query plans in a cache in the context.

The key to these cache maps is the (embedded N1QL query statement + query_context) as a string.

If the same query is seen again, we use a cached query plan and execution tree.

Change-Id: I5fc784def1cc78b01de7c8084a095198000bd892
Reviewed-on: https://review.couchbase.org/c/query/+/188833
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>