Refresh unencrypted local connection pool.
MB-35984 excluded recreation of a localhost unencypted connection pool
on TLS config changes.  This isn't correct if (for example) the non-SSL
ports are disabled (strict mode) when the pool should then be discarded.

The worse-case side effect of this is recreation of the pool for
unencrypted localhost connections when TLS is disabled.

Change-Id: Idc7eeabeb49fd25af3cc24659459d57061bf1b27
Reviewed-on: https://review.couchbase.org/c/query/+/188836
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>