MB-56167 Copy optimizer for join enumeration when executing subqueries
For join enumeration, a new optimizer structure is required. When
we start a Server we make a new optimizer and keep using that for
the entire query, including subqueries.

The issue with this defect is when we have multiple arms of
UNION ALL, all the arms are executing in parallel, and if we have
CTE defined in each UNION ALL arm, all the parallel executions
end up using the same optimizer and thus the same builder, which
causes issues since the parallel execution overwrites each other,
which results in random panics.

The fix is to add a Copy() method for the Optimizer interface,
and use a new copy of the optimizer for each subquery execution.

Change-Id: I90696e5f45dea866be52223677082dcaf69f40ec
Reviewed-on: https://review.couchbase.org/c/query/+/188862
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>