MB-56250 Use separate worker thread to remove obsolete index entries from dictionary
When an index is dropped, its entry from dictionary needs to
be removed (from both in-memory cache and on-disk file). The
current mechanism to achieve this is to get a list of valid
indexes and compare with the index entries in the dictionary.
The ones not in current active index list is presumed to be
dropped and their entries removed.

The issue is when removing entries from on-disk files, we
first need to get a handle of the N1QL_CBO_STATS collection
which needs a namespace lock, which is already obtained
earlier in the stack when getting list of indexes.

To fix this, we now use a separate worker thread for removing
obsolete index entries. The keyspace name and index list is
added to a task map (keyed by keyspace name). The worker
thread reads entries from this task map and perform the task
of compare/remove from dictionary. Since this is done in a
separate worker go thread, it does not deadlock with itself.

The grace period is increased, since previously the task is
synchronous, and now it is asynchronous, which means it is
in "checking" state a lot shorter.

Change-Id: I13cf2fcdc722bedaf5f9ca3d4ded068bde32a6e3
Reviewed-on: https://review.couchbase.org/c/query/+/189048
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>