MB-56265 Distinguish two covering indexes with multiple criteria
Currently with CBO, if we have two covering indexes, we compare
pushdown property followed by cost. In case the two indexes are
similar (e.g. same set of index keys with different ordering)
it is possible to have the same pushdown property and cost. In
such cases we just pick one randomly.

In case of non-CBO the planner uses a few other criteria such
as sumKeys and minKeys for comparing two indexes. Add more
criteria in case of CBO as well, such that in case of same
pushdown property and same cost, we also compare (in order)
cardinality, then sumKeys, then minKeys. The later criteria is
only considered when the earlier criteria are the same. In
most cases we'll still pick by the earlier criteria like
pushdown property and cost.

Change-Id: Id0bec068b86c86bb4e06d467391e42d69ffce711
Reviewed-on: https://review.couchbase.org/c/query/+/189104
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>