MB-55463 Return StatusOK for transient GSI errors
GSI has requested we return 200 rather than 202 for GSI transient errors
as SDKs (etc.) don't know what to do with 202.  I have highlighted we'll
probably get some push-back for 200 too as indices aren't immediately
ready and so a query submitted straight afterwards (typically this is
something like BUILD INDEX that is returning the transient error) will
use the wrong access path or will fail, since SDKs (etc.) will not be
aware they have to wait (given they seem to only care for the HTTP
status and not the warnings returned).

Change-Id: Ibb90d51fc27e5c331c9c52ee46c8603e07c2be49
Reviewed-on: https://review.couchbase.org/c/query/+/189127
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>