MB-56281 Modify OR-clause handling
This defect is a regression from two previous changes
relating to OR-clause handling.

The fix for MB-55236 changed how we account for
sargLength with OR clause, and regard an index key
to be sargable only when it is sargable for all OR
subterms. This is incorrect, if one subterm of OR
can sarg 3 keys and another subterm only sargs 2
keys, we should use 3 as sarg keys, otherwise we'll
not be able to use the 3rd index key.

Another issue is from one of the fixes for
MB-51817/MB-52253. For OR predicates we used to
always call sargForOr and use the individual subterms
for sarging indexes. We changed this to use individual
filters to sarg the indexes first. This works if the
OR-clause is "uniform" like (c1 = 1 OR c1 = 2),
however if the OR clause involves different fields
this change may limit the number of index keys that
can be sarged.

The fix involves several aspects:

 - partially reverse the changes for MB-55236. In
   case of OR-clause, treat an index key as sargable
   if it is sargable for at least one subterm of OR
 - when comparing UnionScan with a single index scan
   (with OR clause) use maxSargLength for all subterms
   of OR, instead of minSargLength. This is consistent
   with the above change since we now regard an index
   key as sargable if it is sargable in any OR subterm
 - when sarging index keys with an OR clause, first
   check whether each subterm of the OR sargs the same
   set of index keys, if so, allow the change for
   MB-51817 to continue, otherwise, revert back to
   previous behavior and sarg each subterm of OR
   individually. This allows different number of index
   keys to be sarged for different OR subterms.
 - a separate issue is exposed while investigating:
   the behavior of MB-42039/MB-42187 resurfaced
   (after fixing MB-54568) in that we use intersect scan
   on two equivalent indexes. Modify the conditions at
   the end of function narrowerOrEquivalent, such that
   after the other conditions are checked, we try to
   differentiate by length of index keys, followed by
   checking for primary index vs missing-leading-key
   index, followed by non-nil index condition, and
   failing all that, check whether the two indexes are
   equivalent. This ensures that we pick one of the
   equivalent indexes rather than using both of them.

Change-Id: I9ecbf29ff9d03ddfb9f1655b7c58b443c0df86f6
Reviewed-on: https://review.couchbase.org/c/query/+/188425
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/188427
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/189184