MB-56276 Add buffer to runQueue
Adding a buffer equal to the number of servicers to a queue allows for a
temporary stretch in the queue to allow for additions between queue
count reduction and actual request release.  This buffer prevents
wrapping overwrite with a full queue.

Pushed the head & tail indicators to 64-bits since we can with this and
GOQUERY_GOMAXPROCS/GOMAXPROCS (and technically always could) have a
non-power of 2 queue size which would not play nicely with wrapping the
32-bit number.  The 64-bit numbers shouldn't wrap during any practical
uptime (thousands of years at 200k requests/sec).

Also enhanced the request scheduler to check the contents of an
expectedly idle queue for stragglers that somehow haven't been scheduled
to run, and run them.  This is a fail-safe and (hopefully) will never be
invoked.

Change-Id: I7c954c997e5e109d9c95d1e803c0d2953b28b442
Reviewed-on: https://review.couchbase.org/c/query/+/189235
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>