MB-56318 Ensure bucket updater is stopped on bucket close
Have streaming function directly respond with the need to exit the
updater rather than indicating this via the StopUpdater function.  This
will ensure that even if a new updater has been kicked off, the existing
updater will exit.

Ensure bucket close stops updater and that updater is stopped when pool
is closed.

Don't use loadNamespace as it duplicates the namespace and buckets
(which can lead to updaters leaking).

Also enhanced bucket updater messages so we can more easily tie messages
to an individual bucket object instance.

Change-Id: I4808043faab678df6cb4af61e28e76e1b9bcb8e6
Reviewed-on: https://review.couchbase.org/c/query/+/189348
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>