MB-56334 Properly detect MISSING/NULL value in hashed fields in hash table
The hash table code relies on value comparisons to determine whether
two values are the same. This logic is used when the hash table is
being built (in case of hash value collision) as well as when the
hash table is being probed.

The issue is that when the hash table is being built, if we have
multiple MISSING/NULL values, since value comparison does not return
true for comparing two MISSING or two NULL values, we treat them as
different values and put each MISSING/NULL value in a different slot
in the hash table. This is not wrong but inefficient since they
occupy hash table slots unnecessarily. This can happen in case of
composite hash join key (multiple equality predicates) and one of
the keys has a MISSING/NULL value.

This bug does not give wrong results for hash join, it is simply
inefficient, especially if there are large number of MISSING/NULL
values in the build side of a hash join.

This is not an issue if a hash table is being used for IN-list
processing since we check for value type there before we put in the
hash table.

The fix contains:

 - the hash table Put() and Get() calls currently takes a single
   value, a marshal function pointer and an equal function pointer.
   change this to take an interface which could be a single value
   or an array of values ([]interface{}).
 - hash table entry gets an array indicator, to signify whether
   a single value or an array of values are used for hashing key
 - the marshal function now has two flavors, one for single value,
   one for array of values
 - the equal function also has two flavors, one for single value
   and one for array of values
 - a couple of new equal functions are introduced, one for single
   value and one for array of values. They both treat two
   MISSING/NULL values as equal, i.e. when comparing two MISSING
   values or two NULL values, it returns true (this is similar to
   IsNotDistinctFrom())
 - hash join/nest uses array if number of join keys is more than
   one, and single value if only one join key. In-list processing
   uses single value
 - when building the hash table, use the new equal function that
   treats two MISSING/NULL values as equal, such that we properly
   detect duplicated join values
 - when probing the hash table, use the original equal function
   that returns MISSING or NULL (which is false) when comparing
   MISSING/NULL values
 - use memory pool for the value arrays used in the hash table
 - users of the hash tables are adjusted for the new behavior

Change-Id: Ib7ba08b2c0a30d25d94df8810c04ff1aee1bba6e
Reviewed-on: https://review.couchbase.org/c/query/+/189353
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>