MB-56068 Set number of CPUs dynamically
Accept "num-cpus" command line argument to override the number of CPUs
seen by the service.  Similarly accept a "num-cpus" setting from
ns_server for the same purpose.  The setting is always limited to the
actual number of logical CPUs known to the process.

If the value provided by either of these mechanisms is greater than zero
then it supercedes all other settings (i.e. GOQUERY_GOMAXPROCS and
GOMAXPROCS). A setting of 0 means (effectively) "not set" so the default
processing - i.e. environment variables or number of logical CPUs with
the 80% limit for serverless - applies.

Changing the number of CPUs after start-up does NOT change the number of
servicers & plus servicers.  To have these be recalculated, set them to
0 after the number of CPUs has been successfully changed.  The request
queue sizes are never altered.

Change-Id: I7673c3b489f81c2d00d0e25e404f3b7df8304892
Reviewed-on: https://review.couchbase.org/c/query/+/189499
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>