MB-56314 Implement a partial graceful shutdown
And make it the default.  In this mode a node will wait for existing
requests and transactions (i.e. those started before a shut down
request) and accept new requests and transactions but will not wait for
them to complete (this aspect is like pre-MB-40533 behaviour).  Once all
the monitored requests and transactions are complete the node reports
"shut down" on the internal endpoint allowing a rebalance to progress
and it to be killed, regardless of activity that was started after the
shutdown was initiated.

A SIGTERM will initiate a shutdown and when complete, the process will
exit.

Change-Id: Ic9bd54903fb34cab93676fcffc89e605392beb80
Reviewed-on: https://review.couchbase.org/c/query/+/189531
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>