MB-56314 Implement a partial graceful shutdown
And make it the default.  In this mode a node will wait for existing
requests and transactions (i.e. those started before a shut down
request) and accept new requests and transactions but will not wait for
them to complete (this aspect is like pre-MB-40533 behaviour).  Once all
the monitored requests and transactions are complete the node reports
"shut down" on the internal endpoint allowing a rebalance to progress
and it to be killed, regardless of activity that was started after the
shutdown was initiated.

A SIGTERM will initiate a shutdown and when complete, the process will
exit.

Change-Id: I4987cbe9df40982ed5bc93630fb20c5cbcb99489
Reviewed-on: https://review.couchbase.org/c/query/+/189568
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>