MB-56462 Retain expression error context with formalization
For queries involving a single keyspace we allow field names
to be directly specified without being prefixed with keyspace
alias, then internally the formalization process will add the
necessary prefix. However the error context (generated from
the parser) is lost during this formalization process.

The fix is to inherit the error context information during
formalization. (expression flags is are bit trickier and it
is not clear whether we should be blindly inheriting all
expression flags. For now we'll leave it out and if there
are specific flags we identify in the future that need to be
inherited we'll do it then).

Change-Id: I1857995c211f0b7981bd30e8ee02750b6ac6c719
Reviewed-on: https://review.couchbase.org/c/query/+/189679
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>