MB-56431 Detect conflicts with WITH aliases during formalization
Prior to the fix for MB-44279, WITH aliases are pushed to
the formalizer in the beginning of a SELECT statement as
a binding variable. This prevents any other usage of the
WITH alias in the query, e.g. an alias used in the FROM
clause, or as a binding variable name. One issue with
this approach is that we cannot effectively check the
usage of an WITH alias as an ExpressionTerm (FROM clause)
as indicated by MB-44279.

The fix for MB-44279 changed the formalization for WITH
aliases. We now treat it similar as any other FROM term
alias, i.e. if the WITH alias is used in the FROM term,
at that time we push it as an allowed alias, otherwise,
any other references to the WITH alias will treat it as
a general expression reference.

One side effect of the fix for MB-44279 is that we are
no longer checking for duplicated references to the WITH
alias, i.e. we now allow for example the usage of a WITH
alias as a alias for a FROM term (keyspace, expression,
subquery), or as a binding variable. Although one can
argue that this behavior is acceptable, which effectively
means that we are overriding the WITH alias with the new
usage, this does create potential confusion, and is a
different behavior than previous releases.

Therefore we decide to add back the check for duplicated
WITH alias references, the fix includes:
 - add a new member function AllowedAlias() to formalizer
   that checks for allowed alias, with two extra boolean
   arguments:
   - checkWith: whether to include WITH alias in check
   - curScope: whether current scope only (i.e. exclude
     parent scope)
 - add a new member function HasAlias() to formalizer
   that checks for existing alias
 - remove current member functions Allowed() and Aliases()
   from formalizer, caller must use one of the above new
   member functions
 - in situations where we want to include WITH alias in
   duplication check, e.g. FROM clause aliases, use
   AllowAlias() checkWith = true and curScope = false,
   this will be majority of situations
 - in situations where we want to exclude WITH alias in
   duplication check, e.g. when handling ExpressionTerm
   where WITH alias is checked separately, use AllowAlias()
   with checkWith = false and curScope = false
 - in situations where we want to check against the
   current scope only (and exclude WITH alias), e.g.
   when checking the FOR clause in index join/nest
   (JOIN <term> ON KEY <key> FOR <keyspace>), use
   AllowAlias() with checkWith = false and curScope = true
 - check for existence of an alias is currently only
   used in ExpressionTerm, and now it uses HasAlias()
 - check for binding variable is handled specially since
   it is in the formalizer code itself, where we just
   add an additional check for WITH aliases
 - fix an existing bug with handling of set operations
   (UNION/EXCEPT/INTERSECT) where we incorrectly set
   projection term alias to be a keyspace alias

Change-Id: Ia5d79f23e97fc961c3631f34c19a1f570a3069ca
Reviewed-on: https://review.couchbase.org/c/query/+/189770
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>