MB-56372 : [UDF profile timings] Check if the cached query plan is still valid and prepare it again if it is not
Also make necessary checks to not use execution trees built using this now invalid query plan.

Implementation:

Every cached plan entry has a "planId" field that is a unique integer.

Similarly, every cached exec tree, has a field "planId" corresponding to the plan that was used to build it.

When checking if the cached exec tree should be used for a statement's execution:
1. Compare the latest cached plan's "planId" of the statement and the "planId" of the tree.
2. If they do not match, then we do not use this tree - and create a new one.

Change-Id: Iab1ee62d9ef4f7aea0841c70e2d44c5dbe14ca19
Reviewed-on: https://review.couchbase.org/c/query/+/189801
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>