MB-56528 Ensure multiple levels of UNNEST dependency are properly handled in ADVISE
The problematic query has 3 levels of UNNEST, and during construction
of unnest map, there is logic to add the "dependency" of an UNNEST to
the unnest map, however currently this logic only handles the current
dependency, it does not handle nested dependency. In case of this bug,
with 3 levels of UNNEST we have two levels of dependencies.

The fix is to make sure we also add nested dependencies.

Change-Id: I390af0ee54cc019b949cb9a754c8fbe5dbcaf469
Reviewed-on: https://review.couchbase.org/c/query/+/189885
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>