MB-56563 Proper setting of exact spans
The symptom of this bug was due to one of the changes of
MB-52253, specifically OR-clause handling. The fix of
MB-52253 was marked as in 7.1.2 although the actual fix
is in 7.1.1 as well.

The change of OR-clause handling is not an issue by
itself for this defect, but it exposed an existing bug
in the setting of exact spans when we sarg index spans
(since a different code path is used now).
Specifically, when we compose index spans for multiple
index keys in a composite index, in function
composeSargSpan(), it is possible that exact is set to
false during processing of the individual spans, e.g.
when query parameters is involved when handling multiple
spans for the same index key. At the end of the function
we should be returning the exact information from the
composed span, instead of the local variable exactSpan.

A separate bug was also identified while investigating.
We are also incorrectly setting exact information when
composing multiple spans for composite index keys in a
different place, where we are currently taking OR of
exact from the two parts, we should be taking an AND.

Once the above bug is fixed, we then have an issue with
the fix for MB-47220 where the symptom resurfaced. We
need to enhance the previous fix such that the index
span is still marked as exact, and we can keep covering
the ANY clause for that bug.

Change-Id: I37372471694f7190391b7e76ec72f7302c41e17d
Reviewed-on: https://review.couchbase.org/c/query/+/189957
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/191243
Well-Formed: Restriction Checker