MB-56068 Double queue buffer
Since we can adjust the number of servicers to a value greater than that
used to create the queues, we could exhaust the buffer added in
MB-56276.  To mitigate this possibility we double the buffer size.

We can still adjust well beyond this initial buffer size and therefore in
theory hit the issue MB-56276 addresses, so we will log a warning when
adjusting beyond the initial servicers count.

Addressed under this MB since dynamic CPU limits are more likely to result
in an upward adjustment of the number of servicers.

Change-Id: Ie3f9ee6884b4464539e412f8ad02c3150b82eeff
Reviewed-on: https://review.couchbase.org/c/query/+/189975
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>