MB-56562 Apply ANY filter on index scan for covering index scan with flatten keys
For array index with FLATTEN_KEYS, the covering is done differently,
where we project the array index key, and apply a modified version
of the ANY clause after the index scan to make sure there is no
false positives. The issue with this bug is that the ANY clause is
applied on a Filter operator, which is after a DistinctScan on top
of the IndexScan3. The DistinctScan only keeps the first entry for
each document, and thus if the index scan is not exact on the array
index key, we may keep a false positive array entry and throw away
an entry that actually satisifies the ANY clause, which may result
in wrong result.

We have mechanism to extract and apply filters for a covering
array index scan already, however this is currently only done when
there are more than one keyspaces in the query. The idea there is
that if there is a single keyspace, then the same filter would
have been applied at the Filter operator after the index scan
anyway (there is no Fetch since it's a covering index scan).

The fix is to enable the extraction and apply of filter on the
index scan itself in case of implicitAny.

Change-Id: I8ff3a42d4487f208f444d44f7edc128c1fccfe6b
Reviewed-on: https://review.couchbase.org/c/query/+/189997
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>