MB-56616 Push filters on unnest alias to covered Unnest Scan
We have a mechanism to push filters that can be evaluated on
the index keys to the index scan itself (early index filter).
However this mechanism currently does not work for the unnest
alias if an Unnest Scan is used (with an array index), since
currently we only look for filters on the base keyspace.

Modify the mechansim to gather index filters to also include
unnest aliases if Unnest Scan is being used. For now this is
specific to covering index scans. This also applies to array
index keys with FLATTEN_KEYS as well.

Change-Id: I7c086c1f3bb76c1ea8018c946710df6a3c522de4
Reviewed-on: https://review.couchbase.org/c/query/+/190131
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>