MB-56616 Push filters on unnest alias to non-covered Unnest Scan
This is the second part of the change for pushing filters on unnest
alias to Unnest Scan. This change is for non-covered Unnest Scan.
We use the same mechanism for early-index filter, except we now
check for filters on unnest aliases as well. For Unnest Scan the
array index key is already replaced with the "unnested" version of
the index key which helps in our checking. Also we don't use
individual filters for sarging unnest index scans, thus we reply
on the same mechanism for OR-clause to determine whether an
individual filter generates an "exact" index span (thus no need
to put in index filter).

Change-Id: I8b4bbf81f2f0a0b263c5479ef79ba70a8ceb08b7
Reviewed-on: https://review.couchbase.org/c/query/+/190255
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>