MB-56372 : [UDF profile timings] Added a revision to the cases we re-use cached plan and execution tree entries.
When the query statement has any of these properties:

1. is inside a transaction
2. has named/positional arguments
3. uses a plan in the prepareds cache

Made the following changes:

UDF statement plan cache :
Do not use any cached entries for the statement.
Do not cache the created plan.

UDF statement exec tree cache:
Do not use any cached entries for the statement
Cache the created execution tree for profiling purposes - but mark the exec tree entry as non-reusable

This is because,  the cached plan entry, might be stale, no longer correct, or sub-optimal. And hence, its corresponding execution tree should not be used.

For example -

If the same statement using named/ positional params is called again in the UDF, we would end up using an earlier cached plan.
This could be incorrect, if the required query plan is different from the cached one - based on the values of the parameters.

Change-Id: If8b02ea20422442de13996e52b83e3155038d280
Reviewed-on: https://review.couchbase.org/c/query/+/190517
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>